Update ScriptHubImporter.py
Adding the ability for the python script to accept a command-line argument for the imports Platform and add a new global variable to the IC_GameVersion64_Import.ahk file. This - when the About addon in Script Hub is updated (pull request incoming) - it can display the platform the imports have been generated for.
If no command-line argument is sent - it doesn't add it.

There may be a better way of doing this - but for the way I generate imports - it felt like the easiest option.